The first working version of worktree detection (#301)
